Skip box conversion if no boxes are present
The `box_convert` function from torchvision expects the input to be a
Tensor[N, 4], where N > 0. Should N == 0 and in_fmt != out_fmt, `unbind`
will error out on the boxes tensor during the conversion process.

The workaround is therefore to skip the box conversion if boxes is an
empty tensor.